requirements.txt
dnspython>=2.4.2
numpy>=1.24.0
typing>=3.7.4.3